refactor(xtask): rename performance experiment harness to perf
