+++ xtask/src/main.rs
@@ -1,8 +1,8 @@
 use anyhow::Result;
 use clap::{Parser, Subcommand};
 use xshell::{cmd, Shell};
-use xtask::bench;
-use xtask::bench::measure::PeakAlloc;
+use xtask::perf;
+use xtask::perf::measure::PeakAlloc;
 
 mod bootstrap;
 mod config;
@@ -27,13 +27,14 @@ enum Commands {
     /// Run e2e tests
     E2eTest,
 
-    /// Storage and minibf benchmarks with shared measurement and reporting
+    /// Storage and minibf performance experiments with shared measurement and
+    /// reporting
     #[command(subcommand)]
-    Bench(bench::Cmd),
+    Perf(perf::Cmd),
 
-    /// Compatibility entry point for storage benchmarks; prefer bench storage
+    /// Compatibility entry point for storage benchmarks; prefer perf storage
     #[command(subcommand)]
-    ArchiveBench(bench::storage::Cmd),
+    ArchiveBench(perf::storage::Cmd),
 
     /// Bootstrap a local Mithril snapshot into an instance
     BootstrapMithrilLocal(bootstrap::BootstrapArgs),
@@ -59,8 +60,8 @@ fn main() -> Result<()> {
             println!("Running sync tests...");
             cmd!(sh, "cargo test --test sync -- --ignored --nocapture").run()?;
         }
-        Commands::Bench(cmd) => bench::run(cmd)?,
-        Commands::ArchiveBench(cmd) => bench::storage::run(cmd)?,
+        Commands::Perf(cmd) => perf::run(cmd)?,
+        Commands::ArchiveBench(cmd) => perf::storage::run(cmd)?,
         Commands::BootstrapMithrilLocal(args) => bootstrap::run(&sh, &args)?,
         Commands::GroundTruth(cmd) => ground_truth::run(cmd)?,
         Commands::TestInstance(cmd) => test_instance::run(&sh, cmd)?,
